requirements.txt
openai
click
python-dotenv
sqlalchemy
pytest
sqlite-vec
